Update installation instructions to clone VVV into directory utilized later in `index.md`.
The existing instructions suggest cloning the VVV repository directly into the current directory (`.`), and direct users to switch to the directory containing vvv (`vagrant-local`) thereafter. In order to make directions directly followable, we correct this inconsistency.

The switch directory step has been retained as users who use the downloadable zip version will be presented with a subdirectory upon extracting the archive. This archive _does not_ match the folder name utilized in the directions; I've chosen to leave this inconsistency as the name that we'd need to use (`VVV-develop`) seems out of place to recommend in the `git clone` instructions.