Transfer documentation from VVV develop branch
The deployment process that we have setup right now for VVV docs
is more complex than it needs to be and involves strange management
of the master/develop branches in VVV separate from that of actual
release cycles.

Moving the documentation into its own repository makes it much
easier to manage and to deploy as needed.

The last state of this documentation in the main VVV repo as well
as its full history can be seen in VVV commit hash 9a930f98d9 or
via GitHub here:

https://github.com/Varying-Vagrant-Vagrants/VVV/tree/9a930f98d9cb2aef84c7bf85c3b6cd1da1faaa58/docs/en-US